no eat robotic

+++ yogstation/code/modules/mob/living/simple_animal/hostile/mining_mobs/marrow_weaver.dm
@@ -64,6 +64,8 @@
 			var/mob/living/carbon/human/H = L
 			var/foundorgans = 0
 			for(var/obj/item/organ/O in H.internal_organs)
+				if(O.status == ORGAN_ROBOTIC)//no tasty organs (lol ipc brains getting eaten)
+					continue
 				if(O.zone == "chest")
 					foundorgans++
 					qdel(O)
